refactor: reorganize features/daily/ root files into forms/, lists/, table/ subdirs + boundary fixes
