+++ src/features/daily/__tests__/BulkDailyRecordForm.test.tsx
@@ -2,7 +2,7 @@ import { TESTIDS } from '@/testids';
 import { cleanup, render, screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { BulkDailyRecordForm } from '../BulkDailyRecordForm';
+import { BulkDailyRecordForm } from '../forms/BulkDailyRecordForm';
 
 // Mock useUsers hook
 const mockUsers = [
