+++ src/features/daily/__tests__/BulkDailyRecordList.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import BulkDailyRecordList, { type BulkDailyRow } from '../BulkDailyRecordList';
+import BulkDailyRecordList, { type BulkDailyRow } from '../lists/BulkDailyRecordList';
 
 const getRowByUserId = (userId: string) => screen.getAllByTestId(`daily-bulk-row-${userId}`)[0];
