+++ src/features/daily/__tests__/TableDailyRecordForm.test.tsx
@@ -2,7 +2,7 @@ import { TESTIDS } from '@/testids';
 import { cleanup, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { TableDailyRecordForm } from '../TableDailyRecordForm';
+import { TableDailyRecordForm } from '../forms/TableDailyRecordForm';
 
 // Mock useUsers hook
 const mockUsers = [
