+++ src/features/daily/__tests__/useTableDailyRecordViewModel.spec.ts
@@ -1,7 +1,7 @@
 import { act, renderHook } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { useTableDailyRecordViewModel } from '../useTableDailyRecordViewModel';
+import { useTableDailyRecordViewModel } from '../table/useTableDailyRecordViewModel';
 
 const navigateMock = vi.fn();
