+++ src/features/daily/domain/dailyRecordLogic.ts
