+++ src/features/daily/forms/TableDailyRecordForm.tsx
@@ -1,8 +1,8 @@
 import { TESTIDS } from '@/testids';
-import { TableDailyRecordHeader } from './components/TableDailyRecordHeader';
-import { TableDailyRecordTable } from './components/TableDailyRecordTable';
-import { TableDailyRecordUserPicker } from './components/TableDailyRecordUserPicker';
-import { useTableDailyRecordForm, type TableDailyRecordData } from './hooks/useTableDailyRecordForm';
+import { TableDailyRecordHeader } from '../components/TableDailyRecordHeader';
+import { TableDailyRecordTable } from '../components/TableDailyRecordTable';
+import { TableDailyRecordUserPicker } from '../components/TableDailyRecordUserPicker';
+import { useTableDailyRecordForm, type TableDailyRecordData } from '../hooks/useTableDailyRecordForm';
 import {
   FilterList as FilterListIcon,
     Group as GroupIcon,
