+++ src/features/daily/lists/BulkDailyRecordList.tsx
@@ -22,7 +22,7 @@ import TextField from '@mui/material/TextField';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 import * as React from 'react';
-import { type MealAmount } from '../../domain/daily/types';
+import { type MealAmount } from '../../../domain/daily/types';
 
 export type BulkRowStatus = 'idle' | 'saved' | 'error' | 'pending';
